provide more memory statistics
* Problem

We need next counters per CT to implement libvirt memory statistics functionality:
swap_in
swap_out
page_in
page_out

* Solution

It is easy to provide "swap_in" and "swap_out" as we have BC counters for them.
"page_in" and "page_out" could be calculated by parsing a lot of /proc files
but as this seems to be too dump these counters are not implemented in this patch.

see #VZOS-60

Signed-off-by: Nikolay Shirokovskiy <nshirokovskiy@odin.com>