configure.ac/Makefiles: remove unused stuff
Some unused artefacts are left in configure.ac since it was taken
from the legacy OpenVZ package. Let's clean this up.

First, a check for libdl is not needed, as the dl library is not used
by this package. Also, remove $(DL_LIBS) from {src,test}/Makefile.am.

Second, remove a line from the configure output that was used to
say that some features were enabled or disabled.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>