Set veid from userspace, #PSBM-34384
We don't want to substitute some legacy ioctls (such as net stats) with cgroup-based interface, because
1) it is difficult to design a good API and
2) since we are unlikely to push it upstream we trade bad for worse by doing it.

So the proposal is to assign an integer ID to each UUID-named container,
which is supposed to be used when working with a legacy ioctl and/or
proc file. IDs will be assigned from userspace before container start
and should meet the following requirements:

1) It must be unique, meaning no two different containers on the same node
may have the same ID. The uniqueness is enforced by the kernel.
2) It must be persistent, meaning the userspace should set ID of a container
to the same value after restart.

It is up to the userspace how to assure the two above mentioned statemensts
always hold. One option is to take first 32 bit of the container's UUID for
its ID, hoping there won't be collisions.

The kernel interface is following:

1) ve.veid is added to each ve cgroup.
2) Initially ve.veid equals 0. Userspace should initialize it by writing an
integer in range (0, INT_MAX] to it. Starting a container without initializing
it will result in ENOENT.
3) If the desired ve.veid is already assigned to another container, write
will return EEXIST. The second and further attempts to set ve.veid will
result in EBUSY unless the value being set equals ve.veid.

Signed-off-by: Igor Sukhih <igor@parallels.com>