scripts/vz-start.in: improve ppp processing
In case of ppp, we need to have a ppp feature enabled in addition
to having access to /dev/ppp, so it's better to check for a feature
rather than for a device.

This helps to avoid false positives like when ppp feature is not set
or disabled inside container.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>