env_nsops: Mount rootfs as a private once cgroups are passed into container
The rootfs should be a private one in a final view, otherwise for systemd
based containers with apache running inside we may reach the situation where
the following picture appear

 | Node
 | ----
 |
 | 53 52 182:616161 / /vz/root/300 rw,relatime shared:31 - ext4 /dev/ploop38510p1 ...
 |
 | Container
 | ---------
 |
 | 66 65 182:616161 / / rw,relatime shared:32 master:31 - ext4 /dev/ploop38510p1 ...
 | 90 66 182:616161 /tmp/systemd-private-Ix7Xyd/tmp /tmp rw,relatime shared:53 master:31 - ext4 /dev/ploop38510p1 ...
 | 91 66 182:616161 /var/tmp/systemd-private-ZzorQo/tmp /var/tmp rw,relatime shared:54 master:31 - ext4 /dev/ploop38510p1 ...

Here root "/" from inside of container is a slave of an external root on the
node, so on restore we're trying to find the "shared:31" and fail simply
because CRIU can't handle such things for now.

The former idea of mounting container's root as a slave came from necessity
of propagation of bindmount cgroup changes. Thus lets a trick for a while
which make both container and criu happy -- mount the root as a slave one,
bindmount cgroups and remount root as a private without recursion.

https://jira.sw.ru/browse/PSBM-34698

CC: Andrew Vagin <avagin@virtuozzo.com>
Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>