env_nsops: Use @pseudosuper feature on the restore procedure
For a long time we've been allowing mounting cgroups from
inside of containers. The primary reason was that we've
had to create venet devices from inside of already existing
VEs (due to its structure).

But we've reworked venet code so now it's aware of net-namespaces.
Still curent design of restore procedure is that we're creating
cgroups from inside of libvzctl on ve0 and then move self
into veX and proceed restore from there.

CRIU in turn restores cgroups remounting them from veX context
(strictly speaking CRIU works in this way to be able to restore
 not only Virtuozzo based containers but general containers as well).

https://jira.sw.ru/browse/PSBM-34299
https://jira.sw.ru/browse/PSBM-43169
https://jira.sw.ru/browse/PSBM-42573

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>
CC: Vladimir Davydov <vdavydov@virtuozzo.com>
CC: Konstantin Khorenko <khorenko@virtuozzo.com>
CC: Andrey Vagin <avagin@virtuozzo.com>
CC: Igor Sukhih <igor@parallels.com>
CC: Pavel Emelyanov <xemul@virtuozzo.com>