CT restore failed if ploop device minor is changed #PSBM-42672
The restore problem rised because device layer in CT is not virtualized.
ploop device name is autogenerated and on CT resume can be changed but
in criu image remains old information.

This patch is workarund the problem by creating device:minor mapping

On CT suspend:
* create mapping in the /dev directoty by creating symlink
uuid -> /dev/pllopN

On CT restore:
* read old device name from uuid
* recreate old device with new major:minor

Signed-off-by: Igor Sukhih <igor@parallels.com>