CT c/r kernel.core_pattern in criu action scripts
We have kernel.core_pattern virtualized per VZCT, that means:
each CT has own sysctl and coredumps are saved relative to VZCT fs
root accoring to that sysctl. Also coredump collector runs in
CTs usermodehelper. We need to c/r these sysctl.

First I tried to do it as criu plugin to save core_pattern in native
criu's protobuf format using pb_read_one pb_write_one, but it turned
out that we can not simply compile in vzct.img protobuf entry for
Virtuozzo's custom properties in CRIU.

On attach to ve/101 cgroup in ve_attach we change task_ve to 101
so it is enouth for sysctl_virtual to work with core_pattern copy
of CT 101.

v2: add vz_ prefix to decrease the chance of images intersection
v3: use cgexec instead of vzctl as this is not good practice to call
vzctl from vzctl hook

https://bugs.openvz.org/browse/OVZ-6696
https://jira.sw.ru/browse/PSBM-44145
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>