Fix ip set on suse CT with wicked. #PSBM-47344
Set of static IP address to interface didn't erase dhcp IP, received earlier.
It was caused by cache of config files in wickedd, which didn't reload after edit of the config file.
We need to restart it to force it read new config files. And do it before any wicked commands to use new data.
If we don't do it wicked will hang or apply merge of old and new configs.

It was done.