scripts: vz -- Handle optional VZ images without error exit
The images we dump with scripts are optional, so we should
not fail if some of them are missing. Moverover in migration
cases not all of them are appearing in constrast with when
c/r cycle is done via vzctl natively.

Thus rework the restore scritps. In particular "action" script
now require any error case must be handled explicitily with
varbose error messaging.

https://jira.sw.ru/browse/PSBM-47359

Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>