Specify quota format on the command line for quotacheck. #PSBM-47292
The 3.13 version of quotacheck, which comes with CentOS 5, contains a bug
in parsing of mount options, and can not correctly determine journaled quota
format from format options. This leads to inability to start the container.

This patch adds a workaround for this issue, explicitly specifying the
quota format on the quotacheck command line.

Also, a new function was introduced which returns the quota format to use,
so it will be easier to add support for other journaled quota formats.

Signed-off-by: Anton Sanarov <asanarov@virtuozzo.com>