Can't stop a container with a lot of nested cgroups #PSBM-49914
It's posible to run the next script in a container and then vzctl will
not be able to stop it and start it again.

CT: cd /sys/fs/cgroup/systemd/; while :; do mkdir a && cd a || break; done
--

If pathlen > pagesize it not possible to remove such dir by full name.
The destroy sysfs directory tree was rewrited to use
  open(O_DIRECTORY)/unlinkat()

Signed-off-by: Igor Sukhih <igor@virtuozzo.com>