scritps/vz-rst-action.in: Fix permissions in restore_devices
When ploop is created on container's startup it's done with
S_IRUSR|S_IWUSR permissions. If container is migrating we need
old device name to be preserved between nodes, for this sake
we create it on restore via action script. But there is a typo:
the device created with 0660 instead of 0600. Fix it.

https://jira.sw.ru/browse/PSBM-57075

Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>