Stop logging before execing function in the container PSBM-58622
If a logger function will be called from an execed function it may write to files which are currently opened from inside the container. This may lead to the file corruption (excess data from logger will present).
This patch will set the logger file to NULL in order to prevent such issues.