Make libvzctl thread safe #PSBM-67089
Using fork() in multithreded may introduce the set of problems.

At the moment of the fork(2) call some threads may be in critical sections
of code, doing non-atomic operations protected by mutexes. In the child
process the threads just disappears and left data half-modified.

It looks like use fork()/exec() is only the sane way to avoid the problem

Signed-off-by: Igor Sukhih <igor@virtuozzo.com>