Preconfigured VETH device #VZOS-1287
We want to leverage libvirt infrastructure for configuring network interfaces.
Thus we need to be able to start containers with veth pair created
and configured by upper layer.

By default we create veth pair inside CT and move the HW peer to host,
so if pair are created on host it could be used to switch to precreated mode.

Signed-off-by: Igor Sukhih <igor@virtuozzo.com>