env_create: created root run path in VE#0 context #PSBM-65033
Currently it's created as a part of first CT pid file. And this file is
created in VE#X context. Which in turn leads to "leaked" dentry.
Well, it's not leaked, but rather anoying, because holds VE#X slabs and
visible in /sys like this after CT stop:

/sys/kernel/slab/:atA-0000192/cgroup/dentry(5:102)

This patch fixes it by explicitly created run root dentry in VE#0

Signed-off-by: Igor Sukhih <igor@virtuozzo.com>