scripts: Tune up ploop links on post-setup-namespaces stage
The restore of ploop device is a bit tricky: its major/minor
is not a constant but allocated dynamicaly by the kernel
driver. Thus we may have a situation when /dev/ploopXX
points for some obsolete name which we have to resolve
into valid device.

For this sake we save old name in dev/$uuid form and
when mount namespaces are restored but their roots
are not moved we should walk over every root point
and figure out if there ploop device to be adjusted.

This patch requires CRIU to export "CRIU_MNT_NS_ROOTS".

https://jira.sw.ru/browse/PSBM-71861

Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>