Move physical network device into CT #PSBM-82259
when we move a physical network device into a Container (read - from host net namespace to CT net namespace), we do this with help of "ip" command, it notifies kernel that sysfs entry "eth1" is now net namespace specific and is not shown on host, but shown inside a Container.

Now when we try to set up permissions on "eth1" sysfs entry via file /sys/fs/cgroup/ve/CTID/ve.sysfs_permissions, it searches for the sysfs entry which is _not_ net namespace specific, fails to fins it and reports an error.

We don't have to set any vz-specific permissions via ve.sysfs_permissions for sysfs entries which are namespace-specific (when checking the permission in sysfs_sd_permission() if we face an entry with non-zero "s_ns", we just say - OK without further checks for permissions set via ve.sysfs_permissions).

Signed-off-by: Igor Sukhih <igor@virtuozzo.com>