Make cpuX/online non-mandatory sysfs permission #PSBM-103389
According to kernel's Documentation/cpu-hotplug.txt file "online" might be missing for some CPU cores.
It depends on CPU architecture and whether it allows cpuX to be offline. If cpuX is not allowed to ever
go offline, file simply wont exist. Thus, absence of this file should not lead to any failures.

Signed-off-by: Pavel Vokhmyanin <pvokhmyanin@virtuozzo.com>