PSBM-105620 Rename mount directory name to the conventional one
The 'repair' feature in vzctl copies the functionality which is implemented in
Virtuozzo Automator long ago and used there until now. The necessity to
implement the feature on the lower architectural layer is due to new Power
Panel implementation, which doesn't rely on VA anymore.

Thus, for backward compatibility with VA and previous version of Power Panel,
it is better to use '/repair' mount directory - to do not confuse end users and
make the feature re-implementation unnoticed by them.