vz-rst-action/vz8: ignore fs.fsync-enable sysctl if it's not available
We don't (yet) have sysctl fs.fsync-enable on vz8 - in scope of the bug
PSBM-104790 it can be implemented either differently or returned as is,
it is not decided yet. So let's skip it if we have no file without fail
for now.

On vz7 the file should be there so everything would be the same.

https://jira.sw.ru/browse/PSBM-120728

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>