vz-rst-action: remove ve start from scripts
In vz8 kernel we don't have a way to set ve.state="START pid" so only
the init of container can start ve now. So we can't start ve from
scripts now, only from criu code in vz8.

To be able to use the same libvzcl in both vz7 and vz8 (as we currently
do) we need to port this change to vz7 criu too.

So now that both vz7 and vz8 criu start ve by themselfs we can remove ve
start from vzt-rst-action script.

https://jira.sw.ru/browse/PSBM-123162

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>