vzctl2_console_start function now runs under clone VFORK
This is done in order to be able to call this function in a multithreaded
context.
Internaly vzctl2_console_start calls setns to run itself under a mount
namespace of a container's init process. It is impossible to do so from the
main process in case if it has any other processes in the same thread group.

The problem can be described with pthread_create as an example.
pthread_create calls clone(....CLONE_FS), which in the kernel increments
curre->fs->users, which later would lead to EINVAL in setns for a mount
namespace, because it's forbidden for a process to install other mount
namespace when it's current->fs->users != 1

We defeat that by running our code inside of a new temporary process,
created by clone(CLONE_VFORK|CLONE_FILES|CLONE_VM,
where
- CLONE_VFORK is done to block parent process execution until this
process finishes it's work,
- CLONE_FILES shares file descriptor table between the parent and the
child process so that the child could open the master filedescriptor "into"
the fd table of a parent processs.
- CLONE_VM shares the address space to return filedescriptor numbers in
the structure provided by arg.

https://jira.sw.ru/browse/PSBM-130590