In VZ8 add /dev/tty0 with major/minor of a null device
Installation of some of the templates involves running a CT on almost
empty image and installing packages. Packages sometimes require runtime for
configuration. In this case some of the packages might want to make use of /dev/tty*
devices which are not set up anymore on environment creation since the reimplementaion
of vzctl console.
Let's put back one fake /dev/tty* device that is implemented as a /dev/null.
For now let's leave one device only (/dev/tt0), but this can be extended to more
devices and /dev/console as well when we catch a usecase.

https://jira.sw.ru/browse/PSBM-131188