Allow to create a Container if private area directory exists
Use-case is a creation from libvirt, vzct driver. It forces the following
workflow - a user creates hdd first and then "defines" ("create" in vzctl
terms) a Container. Same workflow as it has for QEMU VMs.

In this case a user creates hdd first (by vzpkg create image) and then
vzctl2_env_create is called - and it fails if disk path includes future
VE_PRIVATE. Let's do not fail in this case and re-use existing folder for
Container's creation.