vz-cpt/rst: don't mark cgroups mounts external when have cgroupns
When we have cgroupns (vz8 and vz9), mounting "virtualized" (see
is_virtualized_cgroup() in kernel code) is just allowed in container.
So there is no much point in marking cgroups as external, as they are
now fully internal.

More over if we mark them as external we make criu bind-mount them from
host cgroups, which is cross-userns, and mount flags after this become
locked (see lock_mnt_tree() in kernel). If user (systemd) in container
mounts cgroups interlally it can set any flags it needs. But when we
would restore such mount with cgroup marked as external CRIU would be
restricted by host flags and can fail restoring.

v2: still mark them on restore, if they were not marked on dump, marking
them on restore is a noop, but if we migrate from vz7 and cgroups were
marked on dump we need to also mark them on restore.

https://jira.sw.ru/browse/PSBM-134480
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>