scripts: add support of nfs-server migration
Criu can't dump nfs server, for instance:

- nfsd/rpcpipefs contents is not restored
- opened /proc/<pid>/net/rpc/auth.unix.ip/channel is not restored

This is a workaround to support nfs server migration. Idea is to stop
nfs server before suspending container and restart it after resuming it.
As nfs server is stateless, after restart all nfs clients of this server
would reconect flawlessly, so it should not be much difference between
this w/a and honest migration.

Add timeout for operations with services as if a container is "pause"d
or "snapshot"ed by vzctl or in case container user freezes container
processes we can't stop nfs server, thus just skip this and fail later
as criu would not be able to handle running nfs server. Consider this
yet unsupported.

https://jira.sw.ru/browse/PSBM-138439

Co-Developed-by: Yuriy Vasiliev <yuriy.vasiliev@virtuozzo.com>
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Conflicts:
	scripts/vz-rst.in