cgroup: drop umask when creating cgroup directories
We see that when we create a container it's root cgroups are sometimes
created with 0750 permissions (e.g. from prlctl create) due to creator
umask (0027). This leads later to different behaviour for non-root users in
container comparing to non-root users on host as the former can't create
cgroups in container though the latter can. And we see that systemd in
almalinux-8 CT when creating scope cgroups for a non-root user gets
EACCES due to this problem. This patch fixes the problem by using zero
umask at the time of root cgroup creation.

https://jira.sw.ru/browse/PSBM-142393

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>