umount: send umount event if root mount point is busy #PSBM-141624
Before we used to send event about umount device to early and it brought
some problems with race condition. This fix is trying to umount root after
umountig all subdevices and check if root is BUSY. Only in this case
libvirt sends event and asks other processes to free current device
and try to umount device by using old logic. If mount point is not
busy, it just unmounts it.

Signed-off-by: Aleksandr Leskin <aleksandr.leskin@virtuozzo.com>