vz-rst-action: don't fail the script on non-existing old device
The intention was to skip device file updating if there is no device but
a) due to `set -e` we fail the script if any command returns non-zero
b) || skips the error of left argument in terms of `set -e` thus we
don't fail immediately after detecting that olddev file does not exist
c) return without an argument leads to returning previously encountered
error from the function, which is our "no olddev file" error, and that
finally leads to failing the whole script.

Let's return 0 instead of simple return to fix this problem.

https://jira.sw.ru/browse/PSBM-143037

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>