add if condition for setting and removing ip addr for venet0 #PSBM-143042
After creation, the venet adapter is added to the list of veth devices.
thus, after calling the apply_veth_param function, we may encounter
a situation where the net_if_add script will be called twice. First the script will be called
for venet0 as an adapter, and then venet as an element of the list of veth devices.
We add an additional condition to the loop of all veth netif, so as not to add already
added and remove veth(venet0) if required.
Thus, apply_venet_param is only responsible for adding and editing venet0, and
apply_veth_param for removing venet0.

Signed-off-by: Efim Shevrin <efim.shevrin@virtuozzo.com>