disk: Automount ploop devices #PSBM-140207
Currently when a container's disks are being processed by
vzctl_setup_disk automount is signalled only for dm-based devices. This
results in ploop devices not being automounted when a container is
restarted. Introduce new function is_ploop_device to check that a
device is ploop and also signal autmount for ploop devices as well.

Signed-off-by: Nikolay Borisov <nikolay.borisov@virtuozzo.com>