fs: it is not an error if mount point was unmounted before #PSBM-143602
Bug fix. unmount with error EINVAL means the target is not a mount point.
It happens because we decided to unmount devices earlier in
commit 7f6ba412c8f from 14 Oct 2022
```
umount: send umount event if root mount point is busy #PSBM-141624
```
Error in this place does not allow to us to execute
```
run_action_scripts(h, VZCTL_ACTION_POST_UMOUNT);
```

Signed-off-by: Aleksandr Leskin <aleksandr.leskin@virtuozzo.com>