Create vnet for when vnet0 ips didn't set #PSBM-143587
If the venet0 interface has no ip addresses, the veth pair will not be created.
The absence of a veth device leads to the failure of other scripts, in particular vz-set_rate add
at the moment of creating nft rules.
We extend the apply_venet_param function to create venet0 if no ip has been set.

Signed-off-by: Efim Shevrin <efim.shevrin@virtuozzo.com>