scripts: restore previous functionality for resolvconf #PSBM-145766
This fix reverts commits:
052762b80a6 and ecf7149870320
about using systemctl for checking status of resolvconf and
sytsemd-resolved services. We can not use it when CT is in 'starting'
state.
To avoid problem in debian 11 about existed file
'/etc/systemd/system/dbus-org.freedesktop.resolve1.service'
we append checking of symbolic link
'/etc/systemd/system/systemd-resolved.service' on '/dev/null'
that means it is disabled.
Also we need to update 'resolvconf -u' after CT was started.
For some reason we do it manually for debian 11.

Signed-off-by: Aleksandr Leskin <aleksandr.leskin@virtuozzo.com>