Add a loop when adding an adapter to the firewall zone #PSBM-146965
When adding a network adapter to the firewall zone, it is possible
that the firewall-cmd command will not receive a response from
DBus and will fail with an error, while the network interface
will be added to the trusted zone firewall.
We add a loop in order to repeat the command several
times with some time interval.

Co-authored-by: Artem Vasiliev <artem.vasiliev@virtuozzo.com>

Signed-off-by: Fima Shevrin <efim.shevrin@virtuozzo.com>