env_nsops: use memsw right as memory plus swap
This is a fix for cgroup-v1 case where cur_ms is assigned to x, as
cur_ms is mem+swap and we want to assign swap at this line, we need to
subtract cur_mem from it.

Note: I doubt that it can happen in real life due to mm_check_param
checks that basic ub limits are set. So I fix it just to decrease the
level of confusion in code.

https://jira.vzint.dev/browse/PSBM-149975
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>